aggregator/requirements.txt
requests==2.27.1
rsa==4.8
msgpack==1.0.3
numpy==1.22.3
pycrypto==2.6.1
pdoc3==0.10.0
